tests/wpt/tests/css/css-grid/grid-lanes/subgrid/grid-subgridded-to-grid-lanes/track-sizing/column-subgrid-auto-repeat-auto-004.html

<!DOCTYPE html>
<html>
<title>CSS Grid Lanes Test: intrinsic auto-repeat column lane sized by multiple subgrid items</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3/#subgrids">
<link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.grid-lanes {
    display: grid-lanes;
    flow-tolerance: 0;
    grid-template-columns: repeat(auto-fill, auto);
    width: 600px;
    height: 200px;
}

.grid-lanes > div {
    width: 100px;
    height: 100px;
}

.subgrid {
    display: grid;
    grid-template-columns: subgrid;
    grid-auto-rows: 50px;
}
</style>
<div class="grid-lanes">
  <div></div>
  <div class="subgrid" style="grid-column: auto / span 1;">
    <div style="width: 100px;"></div>
    <div style="width: 130px;"></div>
  </div>
</div>
<script>
  test(function() {
    const container = document.querySelector('.grid-lanes');
    const computedStyle = window.getComputedStyle(container);
    assert_equals(computedStyle.getPropertyValue('grid-template-columns'),
      "150px 150px 150px 150px");
  });
</script>
</html>
